requirements.txt
streamlit
torch
pillow
pandas
sqlite3